Don't assume that all implementations of a virtual method share the same signature
Example:
interface IFace<T> { void Set(T); }
class A : IFace<int> { void Set(int) { ... } }
class B : IFace<T> { void Set(T) { ... } }

Wrong: We find Set(int) first and use that for all of them.

Correct, option 1: Use Set(T) for all of them. It will also work for A
because T will be treated as a wildcard.

Correct, option 2: Use signature of the method that belongs to that class,
i.e. Set(int) for A and Set(T) for the others.

We go with option 2 because it adds less code+complexity.